Add GRUB prefixes, ParseOpts for bootloader state
GRUB measures each command it issues to PCR8. GRUB prefixes this with
context detailing whether it was a GRUB command, kernel commandline, or
module commandline.
Additionally, we add ParseOpts and change the interface for
ParseMachineState. This will allow users to decide whether to parse a
bootloader state if it is known ahead of time and supported (GRUB for
now).