Add HashNonce flag to Attest and VerifyAttestation
Currently, key.Attest takes a slice of data for the attestation nonce.
This gets passed directly to the TPM. However, verifiers that are not
aware of the TPM's max size for extraData can inadvertently send a
challenge that is too large.
This change introduces a HashNonce flag that uses the key's signing
scheme hash algorithm to first hash the AttestOpts nonce. The server
package also introduces the same logic to hash the nonce based on the
Attestation message's AkPub signing scheme.
This allows verifiers to send an arbitrary amount of entropy without
knowing the max hash size of the TPM.